omap: dss2/omapfb: Ensure DISPC IRQ uses updated callback array
DISPC IRQ executes callbacks with dispc.irq_lock released, so there is
a chance that a callback be executed even though it has been unregistered.
Do disable/enable to act as a barrier, and ensure that after returning
from this function, the DISPC IRQ will use an updated callback array,
and NOT its cached copy.

Change-Id: Id9fa04bc95b914ac769a1642610522d6a70df640
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>